Token renewal: Only try if user was logged in before
The difference will be visible, e.g., if a user clicks "Logout", then
logs in to Keycloak independently and revisits demo-ui.
Previously, they would be silently logged in to demo-ui again. Now, they
will see an (instant) redirect through Keycloak in this case.
Keycloak could potentially use this redirect to, e.g., let users choose
which of their accounts they want to use to log in to demo-ui.